Rename require mixin
